refactor(shared): split index.ts into separate modules
- types/task.ts - Task, execution, and validation rule types
- types/sdd.ts - SDD configuration types
- types/mcp.ts - MCP tool and patch types
- types/page.ts - DOM and page structure types
- types/validation.ts - Hallucination check and validation result types
- security/types.ts - Security mode, decision, and approval types
- security/shell-analysis.ts - Shell command analysis logic
- utils.ts - Utility functions and constants
- index.ts - Unified re-exports (public API unchanged)

Closes #23
